+++ gui/demo/commanddemo.cpp
@@ -1,17 +1,18 @@
 #include "commanddemo.h"
-#include "serial/command.h"
+#include "serial/serial.h"
+#include "serial/sectionserial.h"
+#include "serial/canvasserial.h"
 
 CommandDemo::CommandDemo(QWidget* parent, MaestroController* maestro_controller) : SimpleDrawingArea(parent, maestro_controller) {
-	//unsigned int args[] = {90};
-	//Command command = Command(Command::Action::MaestroChangeRefreshInterval, args);
-
-
-	unsigned int args[] = {0, CanvasType::AnimationCanvas};
-	Command command = Command(Command::Action::SectionAddCanvas, args);
 	maestro_controller_->add_section_controller(new Point(10, 10));
+	// TODO
 
+	unsigned char buffer[100] = {0};
+	SectionSerial::add_canvas(buffer, 0, CanvasType::ColorCanvas);
+	Serial::run(maestro_controller_->get_maestro(), buffer);
 
-	Command::run(maestro_controller_->get_maestro(), command.get_command());
+	CanvasSerial::draw_circle(buffer, 0, Colors::GREEN, 5, 5, 2, true);
+	Serial::run(maestro_controller_->get_maestro(), buffer);
 }
 
 CommandDemo::~CommandDemo() {}